@@ -4,31 +4,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.15.0",
-    "@ionic/react": "^5.7.0",
-    "@ionic/react-router": "^5.7.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.21",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-transition-group": "^4.4.2",
-    "framer-motion": "^4.1.17",
-    "gh-pages": "^3.2.3",
-    "material-ui": "^0.20.2",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-modal": "^3.14.3",
-    "react-router-dom": "^5.2.1",
-    "react-scripts": "4.0.3",
+    "@babel/core": "^7.18.6",
+    "framer-motion": "^6.5.1",
+    "gh-pages": "^4.0.0",
+    "node-sass": "^7.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-modal": "^3.15.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.2",
-    "typescript": "^4.4.2",
-    "web-vitals": "^1.1.2"
+    "typescript": "^4.7.4",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -36,14 +23,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest",
-      "plugin:react/recommended"
-    ]
+    "eject": "react-scripts eject",
+    "lint": "npx eslint --fix ."
   },
   "browserslist": {
     "production": [
@@ -58,9 +39,28 @@
     ]
   },
   "devDependencies": {
-    "@types/react-modal": "^3.12.1",
-    "@types/react-router-dom": "^5.1.8",
-    "eslint": "^7.32.0",
-    "prettier": "^2.3.2"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.5",
+    "@types/jest": "^28.1.6",
+    "@types/node": "^18.0.5",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-modal": "^3.13.1",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-transition-group": "^4.4.5",
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^3.2.6",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest-dom": "^4.0.2",
+    "eslint-plugin-jsx-a11y": "^6.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^5.5.1",
+    "prettier": "^2.7.1"
   }
 }
