@@ -1,27 +1,19 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
+
 import "./index.scss";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
-import { ThemeProvider, createTheme } from "@material-ui/core/styles";
+const container = document.getElementById("root");
 
-const theme = createTheme({
-  typography: {
-    fontFamily: "Roboto",
-  },
-});
+console.log(container);
 
-ReactDOM.render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <App />
-    </ThemeProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+if (container) {
+  const root = createRoot(container); // createRoot(container!) if you use TypeScript
+  root.render(<App />);
+}
 
-// If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
