@@ -1,6 +1,7 @@
 import React from "react";
+
 import { IconHolder } from "../Shared/IconHolder/IconHolder";
-import { IonContent, IonHeader } from "@ionic/react";
+
 export const Header = () => {
   // const headerData = {
   //   src: "https://cdn.iconscout.com/icon/premium/png-512-thumb/t-letter-1-840447.png",
@@ -8,8 +9,8 @@ export const Header = () => {
   // };
 
   return (
-    <IonHeader>
+    <div>
       <IconHolder text={"Tannor Franks"} />
-    </IonHeader>
+    </div>
   );
 };
